fix heap remove bug

+++ bigmaac.c
@@ -42,7 +42,7 @@
     tmp->next->previous=tmp->previous; \
     tmp->previous->next=tmp->next; \
 }
-#define BIGMAAC_BUFFERED_SIZE(size) (MAX(size,(unsigned int)(size*2.0)))
+#define BIGMAAC_BUFFERED_SIZE(size) (MAX(size,(unsigned long)(size*2.0)))
 enum memory_use { IN_USE=0, FREE=1 };
 enum load_status { LIBRARY_FAIL=-1,
     NOT_LOADED=0,
